Refactored parallelization test code to move duplicated code into the base parallelization test class. Added two new tests.
