+++ src/test/java/test/thread/parallelization/sample/TestClassIFiveMethodsWithNoDepsSample.java
@@ -1,11 +1,12 @@
-package test.thread.parallelization;
+package test.thread.parallelization.sample;
 
 import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
+import test.thread.parallelization.TestNgRunStateTracker;
 
 import java.util.concurrent.TimeUnit;
 
-public class TestClassGWithNoDepsSample {
+public class TestClassIWithNoDepsSample {
 
     @Parameters({ "suiteName", "testName", "sleepFor" })
     @Test
@@ -60,6 +61,48 @@ public class TestClassGWithNoDepsSample {
                         .setTimeOfEvent(time)
                         .setThread(Thread.currentThread())
                         .addData(TestNgRunStateTracker.EventInfo.METHOD_NAME, "testMethodC")
+                        .addData(TestNgRunStateTracker.EventInfo.CLASS_NAME, getClass().getCanonicalName())
+                        .addData(TestNgRunStateTracker.EventInfo.CLASS_INSTANCE, this)
+                        .addData(TestNgRunStateTracker.EventInfo.TEST_NAME, testName)
+                        .addData(TestNgRunStateTracker.EventInfo.SUITE_NAME, suiteName)
+                        .build()
+        );
+
+        TimeUnit.SECONDS.sleep(Integer.parseInt(sleepFor));
+    }
+
+    @Parameters({ "suiteName", "testName", "sleepFor" })
+    @Test
+    public void testMethodD(String suiteName, String testName, String sleepFor) throws InterruptedException {
+        long time = System.currentTimeMillis();
+
+        TestNgRunStateTracker.logEvent(
+                TestNgRunStateTracker.EventLog.builder()
+                        .setEvent(TestNgRunStateTracker.TestNgRunEvent.TEST_METHOD_EXECUTION)
+                        .setTimeOfEvent(time)
+                        .setThread(Thread.currentThread())
+                        .addData(TestNgRunStateTracker.EventInfo.METHOD_NAME, "testMethodD")
+                        .addData(TestNgRunStateTracker.EventInfo.CLASS_NAME, getClass().getCanonicalName())
+                        .addData(TestNgRunStateTracker.EventInfo.CLASS_INSTANCE, this)
+                        .addData(TestNgRunStateTracker.EventInfo.TEST_NAME, testName)
+                        .addData(TestNgRunStateTracker.EventInfo.SUITE_NAME, suiteName)
+                        .build()
+        );
+
+        TimeUnit.SECONDS.sleep(Integer.parseInt(sleepFor));
+    }
+
+    @Parameters({ "suiteName", "testName", "sleepFor" })
+    @Test
+    public void testMethodE(String suiteName, String testName, String sleepFor) throws InterruptedException {
+        long time = System.currentTimeMillis();
+
+        TestNgRunStateTracker.logEvent(
+                TestNgRunStateTracker.EventLog.builder()
+                        .setEvent(TestNgRunStateTracker.TestNgRunEvent.TEST_METHOD_EXECUTION)
+                        .setTimeOfEvent(time)
+                        .setThread(Thread.currentThread())
+                        .addData(TestNgRunStateTracker.EventInfo.METHOD_NAME, "testMethodE")
                         .addData(TestNgRunStateTracker.EventInfo.CLASS_NAME, getClass().getCanonicalName())
                         .addData(TestNgRunStateTracker.EventInfo.CLASS_INSTANCE, this)
                         .addData(TestNgRunStateTracker.EventInfo.TEST_NAME, testName)
